feat(ui): move sign data type file to component folder
